TO CONVERT Fahrenheit to Celcius Temperature
This HTML page includes an input field where the user can enter a Fahrenheit temperature, a button to trigger the conversion, and a paragraph to display the result. The JavaScript function convertToFahrenheit() takes the Fahrenheit input, converts it to Celsius using the conversion formula (F - 32) * 5/9, and then displays the result on the page.

You can save this code in an HTML file and open it in a web browser to see the converter in action. Just enter a Fahrenheit temperature, click the "Convert" button, and the equivalent Celsius temperature will be displayed.